Added Responsive donut charts
I added Responsive Cards that are that has donut charts and card details beside it. 

Note: the donut chart still does not display legit data since I can't figure out and decided not to touch the controller FRA for the donut charts in FRA page to fetch real data and display them.